git-svn-id: svn://svn.dd-wrt.com/DD-WRT@37940 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/zfs/scripts/zfs.sh
@@ -14,6 +14,7 @@ fi
 PROG=zfs.sh
 VERBOSE="no"
 UNLOAD="no"
+STACK_TRACER="no"
 
 ZED_PIDFILE=${ZED_PIDFILE:-/var/run/zed.pid}
 LDMOD=${LDMOD:-/sbin/modprobe}
@@ -34,7 +35,7 @@ KMOD_ZZSTD=${KMOD_ZZSTD:-zzstd}
 usage() {
 cat << EOF
 USAGE:
-$0 [hvud] [module-options]
+$0 [hvudS] [module-options]
 
 DESCRIPTION:
 	Load/unload the ZFS module stack.
@@ -43,10 +44,11 @@ OPTIONS:
 	-h      Show this message
 	-v      Verbose
 	-u      Unload modules
+	-S      Enable kernel stack tracer
 EOF
 }
 
-while getopts 'hvu' OPTION; do
+while getopts 'hvuS' OPTION; do
 	case $OPTION in
 	h)
 		usage
@@ -57,6 +59,9 @@ while getopts 'hvu' OPTION; do
 		;;
 	u)
 		UNLOAD="yes"
+		;;
+	S)
+		STACK_TRACER="yes"
 		;;
 	?)
 		usage
@@ -193,7 +198,7 @@ stack_clear() {
 	STACK_MAX_SIZE=/sys/kernel/debug/tracing/stack_max_size
 	STACK_TRACER_ENABLED=/proc/sys/kernel/stack_tracer_enabled
 
-	if [ -e "$STACK_MAX_SIZE" ]; then
+	if [ "$STACK_TRACER" = "yes" ] && [ -e "$STACK_MAX_SIZE" ]; then
 		echo 1 >"$STACK_TRACER_ENABLED"
 		echo 0 >"$STACK_MAX_SIZE"
 	fi
